+++ obosa_server/src/main/java/com/ssafy/obosa/service/DeleteProductService.java
@@ -3,13 +3,14 @@ package com.ssafy.obosa.service;
 
 import com.ssafy.obosa.model.common.DefaultRes;
 import com.ssafy.obosa.model.domain.Product;
+import com.ssafy.obosa.model.domain.User;
 import com.ssafy.obosa.model.dto.DeleteProductDto;
 import com.ssafy.obosa.repository.ProductRepository;
 import com.ssafy.obosa.repository.UserRepository;
 import com.ssafy.obosa.service.common.FileService;
 import com.ssafy.obosa.util.ImgHandler;
-import com.ssafy.obosa.util.ResponseMessage;
-import com.ssafy.obosa.util.StatusCode;
+import com.ssafy.obosa.enumeration.ResponseMessage;
+import com.ssafy.obosa.enumeration.StatusCode;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
@@ -31,7 +32,7 @@ public class DeleteProductService {
         this.fileService = fileService;
         this.userRepository = userRepository;
     }
-    public DefaultRes<DeleteProductDto> deleteProduct(DeleteProductDto deleteProductDto)
+    public DefaultRes<DeleteProductDto> deleteProduct(User user, DeleteProductDto deleteProductDto)
     {
         try
         {
@@ -45,6 +46,10 @@ public class DeleteProductService {
             }
 
             Product product = optionalProduct.get();
+            if(product.getUser().getUid()!=user.getUid()){
+                return DefaultRes.res(StatusCode.BAD_REQUEST, ResponseMessage.NOT_PERMISSION_ACCESS);
+            }
+
 
             if(product.getImgCount()>0){
                 ImgHandler.deleteProductImgs(fileService, product);
@@ -60,5 +65,13 @@ public class DeleteProductService {
             return DefaultRes.res(StatusCode.INTERNAL_SERVER_ERROR, ResponseMessage.INTERNAL_SERVER_ERROR);
         }
     }
+//    public void Test(int pid){
+//        try {
+//            long result = productRepository.deleteByPid(pid);
+//            System.out.println(result);
+//        }catch (DataIntegrityViolationException e){
+//            e.printStackTrace();
+//        }
+//    }
 
 }
